backend/requirements.txt
addict==2.4.0
albumentations==2.0.3
colorlog==6.9.0
Flask==3.1.0
Flask_Cors==5.0.0
gunicorn>=23.0.0
ipdb==0.13.13
matplotlib==3.10.0
numpy==2.2.2
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
pycocotools==2.0.8
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.15.1
seaborn==0.13.2
setuptools==75.1.0
strawberry-graphql==0.258.0
supervision==0.25.1
termcolor==2.5.0
timm==1.0.14
torch==2.5.1
torchvision==0.20.1
transformers==4.48.2
Werkzeug==3.1.3
yapf==0.43.0
